# apps/server/src/modules/identity/accounts/registration.controller.ts
import { Body, Controller, HttpCode, Post, UseGuards } from '@nestjs/common';
import { ApiBearerAuth, ApiBody, ApiCreatedResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
import { ApiProblemResponses } from '../../../core/http/api-problem-responses.decorator.js';
import { AuthGuard } from '../../../core/http/auth.guard.js';
import { OwnerGuard } from '../../../core/http/owner.guard.js';
import { Public } from '../../../core/http/public.decorator.js';
import { ZodValidationPipe } from '../../../core/http/zod-validation.pipe.js';
import { SensitiveThrottleGuard } from '../auth/sensitive-throttle.guard.js';
import { type AccountView, AccountViewDto } from './account.view.js';
import {
  type AdminCreateUserBody,
  AdminCreateUserDto,
  type RegisterBody,
  RegisterDto,
} from './registration.dto.js';
import { RegistrationService } from './registration.service.js';

/**
 * Self-service registration (technical.md §8, issue #15). Behaviour depends on
 * `registration.mode`: `open`, `invite` (needs `invitationToken`), or `admin`
 * (`403 identity.registration_closed`).
 */
@ApiTags('Registration')
@Controller('auth/register')
export class RegistrationController {
  constructor(private readonly registration: RegistrationService) {}

  @Post()
  @Public()
  @UseGuards(SensitiveThrottleGuard)
  @HttpCode(201)
  @ApiOperation({ summary: 'Register a new account.' })
  @ApiBody({ type: RegisterDto })
  @ApiCreatedResponse({ type: AccountViewDto })
  @ApiProblemResponses({ auth: false, validation: true, statuses: [403, 409, 429] })
  register(@Body(new ZodValidationPipe(RegisterDto)) body: RegisterBody): Promise<AccountView> {
    return this.registration.register({
      name: body.name,
      email: body.email,
      password: body.password,
      displayName: body.displayName,
      invitationToken: body.invitationToken ?? null,
    });
  }
}

/**
 * Owner-only account creation, used when `registration.mode = admin`
 * (technical.md §8, §16).
 */
@ApiTags('Registration')
@ApiBearerAuth('bearer')
@Controller('admin/users')
@UseGuards(AuthGuard, OwnerGuard)
export class AdminUsersController {
  constructor(private readonly registration: RegistrationService) {}

  @Post()
  @HttpCode(201)
  @ApiOperation({ summary: 'Owner-created account (admin registration mode).' })
  @ApiBody({ type: AdminCreateUserDto })
  @ApiCreatedResponse({ type: AccountViewDto })
  @ApiProblemResponses({ validation: true, statuses: [403, 409] })
  create(
    @Body(new ZodValidationPipe(AdminCreateUserDto)) body: AdminCreateUserBody,
  ): Promise<AccountView> {
    return this.registration.adminCreate(body);
  }
}